+++ plugins/infovis/src/histo/BaseHistogramRenderer2D.cpp
