+++ plugins/infovis/src/histo/BaseHistogramRenderer2D.h
