+++ plugins/infovis/src/histo/TableHistogramRenderer2D.h
