+++ plugins/infovis/src/histo/TextureHistogramRenderer2D.cpp
