+++ plugins/infovis/src/histo/TextureHistogramRenderer2D.h
